# docs/openapi.json
{
  "components": {
    "schemas": {
      "DailyUpsertRequest": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "event": {
            "enum": [
              "WASTE",
              "ADJUST",
              "UNFULFILLED"
            ],
            "title": "Event",
            "type": "string"
          },
          "mode": {
            "default": "replace",
            "description": "\"replace\": set the daily total to exactly qty (idempotent). \"sum\": append qty as an additional delta.",
            "enum": [
              "sum",
              "replace"
            ],
            "title": "Mode",
            "type": "string"
          },
          "note": {
            "default": "",
            "maxLength": 500,
            "title": "Note",
            "type": "string"
          },
          "qty": {
            "description": "Units to set (replace) or add (sum).",
            "minimum": 1.0,
            "title": "Qty",
            "type": "integer"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "date",
          "sku",
          "event",
          "qty"
        ],
        "title": "DailyUpsertRequest",
        "type": "object"
      },
      "DailyUpsertResponse": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "mode": {
            "enum": [
              "sum",
              "replace"
            ],
            "title": "Mode",
            "type": "string"
          },
          "noop": {
            "description": "True when replace mode is called with a qty that already matches the current total.",
            "title": "Noop",
            "type": "boolean"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "qty_delta": {
            "description": "Units actually written to the ledger (0 if noop; negative if replace reduced the total).",
            "title": "Qty Delta",
            "type": "integer"
          },
          "qty_total": {
            "description": "Final running total for (sku, date, event) after this call.",
            "title": "Qty Total",
            "type": "integer"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "date",
          "sku",
          "event",
          "mode",
          "qty_delta",
          "qty_total",
          "note",
          "noop"
        ],
        "title": "DailyUpsertResponse",
        "type": "object"
      },
      "ExceptionRequest": {
        "properties": {
          "client_event_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional UUID supplied by the client for idempotency. If the same client_event_id is received a second time the server replays the stored response with already_recorded=true without writing to the ledger again.",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ],
            "title": "Client Event Id"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "event": {
            "enum": [
              "WASTE",
              "ADJUST",
              "UNFULFILLED"
            ],
            "title": "Event",
            "type": "string"
          },
          "note": {
            "default": "",
            "maxLength": 500,
            "title": "Note",
            "type": "string"
          },
          "qty": {
            "minimum": 1.0,
            "title": "Qty",
            "type": "integer"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "date",
          "sku",
          "event",
          "qty"
        ],
        "title": "ExceptionRequest",
        "type": "object"
      },
      "ExceptionResponse": {
        "properties": {
          "already_recorded": {
            "default": false,
            "title": "Already Recorded",
            "type": "boolean"
          },
          "client_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Event Id"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": "integer"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          }
        },
        "required": [
          "date",
          "sku",
          "event",
          "qty",
          "note"
        ],
        "title": "ExceptionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST backend for **desktop-order-system**: stock queries, EAN lookup, exception logging, and receipt closure.\n\nAll write endpoints require a `Bearer <DOS_API_TOKEN>` header.\nSee [`docs/api_contract.md`](https://github.com/sasu91/desktop-order-system/blob/main/docs/api_contract.md) for the full contract.",
    "title": "desktop-order-system API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/exceptions": {
      "post": {
        "description": "Scrive un evento WASTE, ADJUST o UNFULFILLED nel ledger.\n\n**Idempotenza con client_event_id**\n\nInserisci nel payload un UUID univoco per ogni evento logico:\n\n```json\n{ \"date\": \"2026-02-25\", \"sku\": \"PRD-0042\", \"event\": \"WASTE\",\n  \"qty\": 3, \"client_event_id\": \"550e8400-e29b-41d4-a716-446655440000\" }\n```\n\nSe la stessa richiesta arriva due volte (es. retry di rete), il server\nrisponde **200** con `already_recorded=true` — il ledger non viene\ntoccato una seconda volta.\n\n**Senza client_event_id**: ogni chiamata viene accettata — non si usa\nla tripletta ``date+sku+event`` come chiave legacy (niente 409).",
        "operationId": "create_exception_api_v1_exceptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExceptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            },
            "description": "Già registrato — risposta replay (client_event_id duplicato)"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            },
            "description": "Eccezione registrata con successo"
          },
          "400": {
            "description": "Input non valido"
          },
          "404": {
            "description": "SKU non trovato"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Registra eccezione (WASTE / ADJUST / UNFULFILLED)",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/api/v1/exceptions/daily-upsert": {
      "post": {
        "description": "Mantieni un **unico totale giornaliero** per la tripletta `(sku, date, event)`.\n\n## Modalità\n\n| `mode`      | Semantica |\n|-------------|--------------------------------------------|\n| `\"replace\"` | Imposta il totale a esattamente `qty`. Idempotente: se il totale corrente è già `qty`, risponde `noop=true` senza toccare il ledger. |\n| `\"sum\"`     | Aggiunge `qty` come delta al totale corrente. Ritorna il nuovo totale. |\n\n## Differenza con `POST /exceptions`\n\n`POST /exceptions` **aggiunge sempre** una nuova riga al ledger;\ndue chiamate nella stessa giornata producono due eventi separati\n(corretto per scarti multipli, movimenti distinti, ecc.).\n\n`POST /exceptions/daily-upsert` è progettato per flussi che inviano il\n**totale cumulativo** della giornata (es. integrazione ERP, fine giornata\nPOS) dove duplicare le righe produrrebbe un doppio conteggio.",
        "operationId": "daily_upsert_exception_api_v1_exceptions_daily_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyUpsertResponse"
                }
              }
            },
            "description": "Upsert eseguito (o noop se totale già corretto)"
          },
          "400": {
            "description": "Input non valido"
          },
          "404": {
            "description": "SKU non trovato"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert totale giornaliero (WASTE / ADJUST / UNFULFILLED)",
        "tags": [
          "exceptions"
        ]
      }
    },
...
  }
}